+++ trunk/SipStack/SipICTList.cpp
@@ -249,7 +249,7 @@ void CSipICTList::DeleteAll( )
  */
 int CSipICTList::GetSize( )
 {
-	return m_clsMap.size();
+	return (int)m_clsMap.size();
 }
 
 /**
